Add files via upload
The main objective of this project is to create a user-friendly and visually appealing version of the Tic-tac-toe game. The game will be designed to allow two players to compete against each other on a 3x3 grid. The players will take turns marking their symbols (X or O) on the grid until one player wins or the game ends in a draw. The project aims to provide an engaging and enjoyable gaming experience to users while showcasing the capabilities of Java Swing for GUI development.